+++ src/main/java/org/example/servicelinkbe/configuration/security/token/impl/AccessTokenImpl.java
@@ -12,10 +12,10 @@ import java.util.Set;
 @Getter
 public class AccessTokenImpl implements AccessToken {
     private final String subject;
-    private final Integer userId;
+    private final Long userId;
     private final Set<String> roles;
 
-    public AccessTokenImpl(String subject, Integer userId, Collection<String> roles) {
+    public AccessTokenImpl(String subject, Long userId, Collection<String> roles) {
         this.subject = subject;
         this.userId = userId;
         this.roles = roles != null ? Set.copyOf(roles) : Collections.emptySet();
